server/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
python-decouple
Pillow
drf-spectacular
stripe
django-cors-headers